backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   starlette
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.0
    # via uvicorn
fastapi==0.115.12
    # via backend (pyproject.toml)
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core==2.24.2
    # via google-ai-generativelanguage
google-auth==2.40.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-genai
google-genai==1.15.0
    # via backend (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.2
    # via sqlalchemy
grpcio==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via backend (pyproject.toml)
langchain-core==0.3.59
    # via
    #   langchain
    #   langchain-google-genai
    #   langchain-text-splitters
langchain-google-genai==2.1.4
    # via backend (pyproject.toml)
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.42
    # via
    #   langchain
    #   langchain-core
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
pillow==11.2.1
    # via backend (pyproject.toml)
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.4
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.10
    # via backend (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.4
    # via
    #   fastapi
    #   google-genai
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via backend (pyproject.toml)
pymupdf==1.25.5
    # via backend (pyproject.toml)
python-dotenv==1.1.0
    # via pydantic-settings
python-multipart==0.0.20
    # via backend (pyproject.toml)
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
requests==2.32.3
    # via
    #   google-api-core
    #   google-genai
    #   langchain
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via langchain
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via langchain-core
typing-extensions==4.13.2
    # via
    #   fastapi
    #   google-genai
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via backend (pyproject.toml)
websockets==15.0.1
    # via google-genai
zstandard==0.23.0
    # via langsmith
